+++ app/enums/sources.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class Sources(str, Enum):
+    """
+    A source available for retrieving data.
+    """
+    jhu = 'jhu'
+    csbs = 'csbs'
