+++ app/router/__init__.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+# Create the router.
+router = APIRouter()
+
+# The routes.
+from . import latest, sources, locations
